Extracted chat-scroll action
